% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_arima.R
\name{arimaUnivariate}
\alias{arimaUnivariate}
\title{Auto.Arima Default}
\usage{
arimaUnivariate(df, end_forecast)
}
\arguments{
\item{df}{DataFrame contendo a série limpa e organizada;}

\item{end_forecast}{representa a data final, último mês ou trimestre, da projeção;}
}
\value{
Retorna o mesmo df de input, porém a projeção formalmente aplicada com o método Auto.Arima
}
\description{
Aplicação do método Auto.Arima para projeção
}
\details{
O input deve ser um data.frame organizado:
\code{date}: Data da observação:
\code{vl}: valor da observação.
}
\examples{
\dontrun{

arimaUnivariate(df, end_forecast = "2026-12-01")

}

}
\author{
Luiz Paulo T.
}
