% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_yoy.R
\name{calc_yoy}
\alias{calc_yoy}
\title{Calculate yoy from target value}
\usage{
calc_yoy(df, target_aop)
}
\arguments{
\item{df}{Dataframe contendo a série limpa e organizada;}

\item{target_aop}{Vetor de valores indicando a projeção desejada para média de período;}
}
\value{
Retorna um vetor de valores de mesmo cumprimento de @param target_aop,
porém com os valores representando o YoY para o ano.
}
\description{
Converte um vetor de valores, que estão na mesma unidade de medida da séries, e representam a média de um período futuro
para a série, em um vetor contendo variações YoY entre os anos.
}
\details{
O @param df de entrada deve conter pelo as colunas de:
\code{date}: Data da observação:
\code{vl}: valor da observação;

O @param target_aop indica o valor para média de período desejado, idealmete advindo de uma projeção anual.
Por exemplo, a projeção anual do LatamFocus aponta média de 15% para 2022 e 12% para 2023. Pode-se preencher:
target_aop = c(0.15,0.12)
}
\examples{
\dontrun{
calc_yoy <- function(df,
                     target_aop = c(11.5,10.5,10.1,9.8,9.5))
}

}
\author{
Gabriel Bellé
}
