% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cambio_real.R
\name{cambio_real}
\alias{cambio_real}
\title{Calculate real values exchange rate}
\usage{
cambio_real(df, df_ipc_dom, df_ipc_int, mes_base)
}
\arguments{
\item{df}{Dataframe contendo a série de câmbio limpa e organizada;}

\item{df_ipc_int}{Dataframe contendo a série de inflação da moeda estrangeira;}

\item{mes_base}{Ano e mês base para o câmbio real.}

\item{df_ipca_dom}{Dataframe contendo a série de inflação da moeda doméstica;}
}
\value{
Retorna um dataframe contendo o câmbio real.
}
\description{
A função calcula o câmbio ajustado pela inflação das moedas computadas em sua taxa, câmbio real.
}
\details{
O @param df de entrada deve conter pelo as colunas de:
\code{date}: Data da observação:
\code{vl}: valor da observação;
}
\examples{
\dontrun{
cambio_real(real_usd, #BRL/USD
            ipca_br,
            ipc_us,
            '2022-01-01'
            ))
}

}
\author{
Gabriel Bellé
}
