% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vector_len.R
\name{check_vector_len}
\alias{check_vector_len}
\title{Check input vector length is equal or less than needed}
\usage{
check_vector_len(df_forecast, vector_to_check)
}
\arguments{
\item{df_forecast}{Dataframe contendo a série com o período de projeção calculado;}

\item{vector_to_check}{Vetor contendo o alvo do drift ou os valores do drift manual.}
}
\value{
O retorno é um vetor com o tamanho adequado em relação ao número de meses
da projeção e o número de vezes que cada valor será repetido.
}
\description{
Compara a quantidade de valores no vetor de input com o número necessário
para fazer projeção utilizando o método drift.
}
\details{
O @param df_forecast de entrada deve conter pelo as colunas de:
{date}: Data da observação:
{vl}: valor da observação;
{forecast}: bool indicando se o valor na data é projetado.
}
\examples{
check_vector_len(df_forecast = df_forecast,
                  vector_to_check = target_value)

}
\author{
Gabriel Bellé
}
