% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_hist.R
\name{drift_hist}
\alias{drift_hist}
\title{Drift forecast with historical tendency.}
\usage{
drift_hist(df_forecast, nyears = NULL)
}
\arguments{
\item{df_forecast}{Dataframe contendo a série a ser projetada;}

\item{nyears}{Restrição de anos para uso do histórico da série.}
}
\value{
O retorno é um df, contendo as colunas de:
\code{date}; \code{vl}; \code{drift} e \code{forecast}.
As colunas são iguais ao input, com exceção da coluna drift, que possui o valor a ser adicionado na série.
}
\description{
Incorpora tendência na projeção de uma série.
Para calcular a tendência, utiliza a tendência linear presente no histórico da série.
}
\details{
O @param df_forecast de entrada deve conter pelo as colunas de:
\code{date}: Data da observação:
\code{vl}: valor da observação;
\code{forecast}: bool indicando se a observação é uma projeção.

O @param nyears indica quantos anos do histórico serão utilizados para calcular a tendência linear.
Se nenhum valor fornecido, utilizará o histórico completo.
}
\examples{
\dontrun{
drift_hist(df_forecast = cleaned_df,
           nyears = 5)
}

}
\author{
Gabriel Bellé
}
