% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_manual.R
\name{drift_manual}
\alias{drift_manual}
\title{Drift forecast with pre-defined MoM variation.}
\usage{
drift_manual(df_forecast, manual_drift)
}
\arguments{
\item{df_forecast}{Dataframe contendo a série projetada, a ser adicionada tendência;}

\item{manual_drift}{Vetor de valores numéricos indicando drift em %;}
}
\value{
O retorno é um df, contendo as colunas de:
\code{date}; \code{vl}; \code{drift} e \code{forecast}.
As colunas são iguais ao input, com exceção da coluna drift, que possui o valor a ser adicionado na série.
}
\description{
Incorpora tendência na projeção de uma série.
Para calcular a tendência, utiliza os valores fornecidos, que devem ser % de MoM a ser aplicado na projeção.
}
\details{
O @param df_forecast de entrada deve conter pelo as colunas de:
\code{date}: Data da observação:
\code{vl}: valor da observação.
\code{forecast}: bool indicando se a observação é uma projeção.

O @param manual_drift deve conter um vetor numérico onde cada constante
indica a variação mensal desejada na projeção para cada ano. Caso o tamanho seja menor que
a quantidade de anos a ser projetada, repetirá-se o último valor colocado.
Ex: manual_drift = c(0.1, 0.1, 0.15, 0.2).
Assim, será utilizado 0.1% nos dois primeiros anos, 0.15% no terceiro, 0.2% no quarto e nos anos seguintes,
até o fim do horizonte da projeção.

Atenção: este parâmetro faz com que a tendência seja exponencial. O aumento em t+1 é uma % do valor em t + valor em t.
}
\examples{
\dontrun{
drift_manual(df_forecast = cleaned_df,
             manual_drift = c(0.1, 0.15))
}

}
\author{
Gabriel Bellé
}
