% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_target.R
\name{drift_target}
\alias{drift_target}
\title{Drift forecast with target values}
\usage{
drift_target(df_forecast, target_value, trend_type = NULL)
}
\arguments{
\item{df_forecast}{Dataframe contendo a série a ser projetada;}

\item{target_value}{Vetor de valores indicando a projeção desejada para final de período;}

\item{trend_type}{o valor no parâmetro irá modificar a fórmula empregada para cálculo do drift quando utilizado os valores
alvo em target_value. Aceita os valores (linear, exponencial).}
}
\value{
O retorno é um df, contendo as colunas de:
\code{date}; \code{vl}; \code{drift} e \code{forecast}.
As colunas são iguais ao input, com exceção da coluna drift, que possui o valor a ser adicionado na série.
}
\description{
Incorpora tendência na projeção de uma série.
Para calcular a tendência, utiliza a diferença entre o último dado realizado e o valor alvo para o ano.

Atenção! O comportamento dessa função funciona como esperado apenas quando a projeção de entrada foi
originada de um método naive (onde a projeção é apenas a repetição do último valor).
}
\details{
O @param df_forecast de entrada deve conter pelo as colunas de:
\code{date}: Data da observação:
\code{vl}: valor da observação;
\code{forecast}: bool indicando se a observação é uma projeção.

O @param target_value indica o valor para o final de período desejado, idealmete advindo de uma projeção anual.
Por exemplo, a projeção anual do LatamFocus aponta 150 para 2023 e 200 para 2024, pode-se preencher:
end_projection = '2025-12-01' e target_value = c(150, 200)

Isto fará com que a tendência linear seja tal qual respeite os valores de entrada.
}
\examples{
\dontrun{
drift_target(df_forecast = cleaned_df,
             target_value = c(200, 230, 150, 210, 100))
}

}
\author{
Gabriel Bellé
}
