% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_series.R
\name{expand_series}
\alias{expand_series}
\title{Expand date span in a time series}
\usage{
expand_series(df, end_forecast)
}
\arguments{
\item{df}{DataFrame contendo uma linha com data e observação da série;}

\item{end_forecast}{Date informando quando a projeção encerra.}
}
\value{
Retorna o mesmo df de input, mas com o período de datas aumentado de acordo com
o fim da projeção fornecida.
}
\description{
Aumenta o vetor de date de um dataframe, inputando NAs na coluna de observação,
para ser preenchida por uma função de forecast.
}
\details{
O input deve ser um df contendo pelo as colunas:
\code{date}: Data da observação:
\code{vl}: valor da observação.
}
\examples{
\dontrun{
expand_series(df = df_series,
              end_forecast = '2026-12-01')
}

}
\author{
Gabriel Bellé
}
