% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_periodicity.R
\name{get_periodicity}
\alias{get_periodicity}
\title{Discover periodicity from df}
\usage{
get_periodicity(df)
}
\arguments{
\item{df}{Dataframe contendo a base de dados.}
}
\value{
O retorno é uma lista contendo os parametros de:
{p_name}: nome da periodicidade;
{p_nmonths}: quantos meses há na periodicidade;
{p_ngap}: quantos meses pular até a próxima observação.
}
\description{
Descobre a periodicidade de um dataframe e cria parametros adequados a ela.
}
\details{
O input deve ser um df contendo pelo menos as colunas de:
{date}: Data da observação:
{vl}: valor da observação.
}
\examples{
get_periodicity(df = cleaned_df)

}
\author{
Gabriel Bellé
}
