% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_realvl.R
\name{get_realvl}
\alias{get_realvl}
\title{Calculate real values}
\usage{
get_realvl(df, df_cpi, dt_base_in, dt_base_fim = NULL)
}
\arguments{
\item{df}{Dataframe contendo a série nominal limpa e organizada;}

\item{df_cpi}{Dataframe contendo a série de inflação da moeda;}

\item{dt_base_in}{Ano e mês base de inicio para o indice de inflação.}

\item{dt_base_fim}{Ano e mês base de fim para o indice de inflação.}
}
\value{
Retorna um dataframe contendo a série real.
}
\description{
A função calcula o a deflação de uma série nominal.
}
\details{
Todos dataframes devem começar pelo menos 1 mês antes da data inserida em mes_base.
}
\examples{
get_realvl(price_fuel, #Preço gasolina
           ipca_br,
           '2022-12-01'
           )

}
\author{
Gabriel Bellé
}
