% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seas_adj.R
\name{get_seas_adj}
\alias{get_seas_adj}
\title{Seasonality adjust}
\usage{
get_seas_adj(df, type = "mean")
}
\arguments{
\item{df}{Dataframe contendo a série a ser dessazonalizada;}

\item{type}{Chr especificando tipo de dessaz a ser aplicado (STL, X13, mean).}
}
\value{
O retorno é um df contendo os valores da série dessazonalidada e a data.
}
\description{
Calcula o dessaz de uma série com as configurações automáticas
do STL ou do X13, opcional retornar a média dos dois.
}
\details{
O input @param df deve ser um DataFrame contendo pelo menos as colunas de:
\code{date}: Data da observação:
\code{vl}: valor da observação.
}
\examples{
\dontrun{
get_seas_adj(df = cleaned_df, type = 'median')
}

}
\author{
Gabriel Bellé
}
