% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stl.R
\name{get_stl}
\alias{get_stl}
\title{Seasonal-Trend decomposition using LOESS (STL)}
\usage{
get_stl(df)
}
\arguments{
\item{df}{representa a série para dessazonalização}
}
\value{
O retorno é um df contendo os valores da série dessazonalidada e a data.
}
\description{
Calcula o dessaz de uma série com as configurações automáticas
do Seasonal-Trend decomposition using LOESS (STL).
}
\examples{
get_stl(df = cleaned_df)

}
\author{
Gabriel Bellé
}
