% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_from_yoy.R
\name{get_target_from_yoy}
\alias{get_target_from_yoy}
\title{Get target from YoY}
\usage{
get_target_from_yoy(df, yoy, end_forecast)
}
\arguments{
\item{df}{Dataframe contendo a série limpa e organizada;}

\item{yoy}{Vetor numérico contendo a variação YoY em \% (1\% = 1)}

\item{end_projection}{Data contendo o último mês a ser projetado.}
}
\value{
Retorna o mesmo df de input, porém a projeção formalmente aplicada com o método naïve,
e com a coluna forecast indicando quais observações são projeção.
}
\description{
Calcula um valor alvo, na mesma unidade de medida da série de input,
que respeite o dado de YoY passado como argumento.
}
\details{
O @param df de entrada deve conter pelo as colunas de:
{date}: Data da observação:
{vl}: valor da observação;
}
\examples{
get_target_from_yoy(df = df_cleaned,
                    yoy = (10,12,12,15)
                    end_projection = '2026-12-01')

}
\author{
Gabriel Bellé
}
