% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World - função teste para estrutura}
\usage{
hello(x)
}
\arguments{
\item{x}{---}
}
\value{
O retorno é ---
}
\description{
---
}
\details{
---
}
\examples{
\dontrun{
hello(x)
}

}
\author{
Gabriel Bellé
}
