% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_from_yoy.R
\name{naive}
\alias{naive}
\alias{get_target_from_yoy}
\title{Naive forecast}
\usage{
get_target_from_yoy(df, yoy, end_forecast)
}
\arguments{
\item{df}{Dataframe contendo a série limpa e organizada;}

\item{end_projection}{Data contendo o último mês a ser projetado.}
}
\value{
Retorna o mesmo df de input, porém a projeção formalmente aplicada com o método naïve,
e com a coluna forecast indicando quais observações são projeção.
}
\description{
Aplica o método naive para realizar projeção, cada valor previsto é igual ao último observado.
}
\details{
O @param df de entrada deve conter pelo as colunas de:
\code{date}: Data da observação:
\code{vl}: valor da observação;
}
\examples{
\dontrun{
naive(df = df_cleaned,
      end_projection = '2026-12-01')
}

}
\author{
Gabriel Bellé
}
