% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_aop.R
\name{sf_aop}
\alias{sf_aop}
\title{Average of period forecasting}
\usage{
sf_aop(df, end_forecast, target_aop, is_yoy = F)
}
\arguments{
\item{df}{Dataframe contendo a série limpa e organizada;}

\item{end_forecast}{Data (YYYY-MM-DD) contendo o último mês a ser projetado;}

\item{target_aop}{Vetor de valores indicando a projeção desejada para média de período;}
}
\value{
Retorna o mesmo df de input, porém a projeção formalmente aplicada com o método,
e com a coluna forecast indicando quais observações são projeção.
}
\description{
A projeção é realizada utilizando um input manual, que deve corresponder ao YoY médio para o ano.
A função ajusta o YoY para os meses restantes do ano, levando em conta os dados realizados, para que o YoY médio corresponda ao input.
}
\details{
O @param df de entrada deve conter pelo as colunas de:
{date}: Data da observação:
{vl}: valor da observação;

O @param target_aop indica o valor para média de período desejado, idealmete advindo de uma projeção anual.
Por exemplo, a projeção anual do LatamFocus aponta média de 15% para 2022 e 12% para 2023. Pode-se preencher:
target_aop = c(15,12). Caso a quantidade de valores seja menor que o número de anos a ser projetado, repete-se o último valor.
}
\examples{
sf_aop(df, end_forecast = '2026-12-01',
       target_aop = c(11.5,10.5,10.1,9.8,9.5)
       )

}
\author{
Gabriel Bellé
}
