% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_conversao_cambio.R
\name{sf_conversao_cambio}
\alias{sf_conversao_cambio}
\title{Modify country in exchange rate}
\usage{
sf_conversao_cambio(df, df_target, df_ratio)
}
\arguments{
\item{df}{Dataframe contendo a série limpa e organizada a ser projetada, contendo apenas dados realizados;}

\item{df_target}{Dataframe contendo a taxa de câmbio apenas com projeção;}

\item{df_ratio}{Dataframe contendo apenas a projeção do par incomum em df e df_target}
}
\value{
Retorna um Dataframe contendo a projeção da série df, no mesmo horizente que df_target
}
\description{
Realiza a projeção de um par de câmbio AB, utilizando a projeção dos pares AC e CB.
(Ex.: par a ser projetado: ARS/EUR, há projeção de ARS/USD e USD/EUR)
}
\examples{
sf_conversaro_cambio(df = ars_eur, df_target = ars_usd, df_ratio = usd_eur)

}
\author{
Gabriel Bellé
}
