% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_daily.R
\name{sf_daily}
\alias{sf_daily}
\title{Daily interpolation}
\usage{
sf_daily(target, ...)
}
\arguments{
\item{target}{Dataframe contendo a série limpa e organizada, mensal e com projeção que será o alvo da projeção diária;}

\item{...}{Dataframe, N dfs que serão projetados utilizando a projeção contida em target}
}
\value{
Retorna um dataframe com colunas de {date}, {forecast} e uma
coluna de valor para cada série que foi projetada, na mesma ordem em que foi inputada na função.
}
\description{
Realiza a projeção de uma série diária, com base em uma série mensal já com projeção.
O último dia da série diária será igual ao valor da série mensal fornecida.
}
\details{
A interpolação feita na projeção é a linear, baseada nas funções 'naive' e 'drift_target'.
A projeção da série diária considera apenas os dias úteis, seguindo calendário da Anbima.
}
\examples{
\dontrun{
sf_daily(target = cambio_real_mensal,
         cambio_diario_venda,
         cambio_diario_compra)
}

}
\author{
Gabriel Bellé
}
