% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_hw.R
\name{sf_hw}
\alias{sf_hw}
\title{Holt-Winters}
\usage{
sf_hw(df, type = NULL, end_forecast)
}
\arguments{
\item{df}{DataFrame contendo a série limpa e organizada;}

\item{type}{Tipo de Holt-Winters: multiplicative, additive, trend e NULL (default);}

\item{end_forecast}{representa a data final, último mês ou trimestre, da projeção.}
}
\value{
Retorna o mesmo df de input, porém a projeção formalmente aplicada com o método Holt-Winters
}
\description{
Aplicação do método Holt-Winters para projeção
}
\examples{
sf_hw(df, type = NULL, end_forecast = "2026-12-01")

}
\author{
Luiz Paulo T.
}
