% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_naive.R
\name{sf_naive}
\alias{sf_naive}
\title{Naive forecast}
\usage{
sf_naive(df, end_forecast)
}
\arguments{
\item{df}{Dataframe contendo a série limpa e organizada;}

\item{end_forecast}{Chr data contendo o último mês a ser projetado.}
}
\value{
Retorna o mesmo df de input, porém com a projeção naïve,
e com a coluna forecast indicando quais observações são projetadas.
}
\description{
Aplica o método naive para realizar projeção. Cada valor previsto é igual ao último observado.
}
\details{
O @param df de entrada deve conter, pelo menos, as colunas de:
\code{date}: Data da observação:
\code{vl}: valor da observação;
}
\examples{
\dontrun{
sf_naive(df = df_cleaned, end_forecast = '2026-12-01')
}

}
\author{
Gabriel Bellé
}
