% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_seas_ratio.R
\name{sf_seas_ratio}
\alias{sf_seas_ratio}
\title{Seasonality ratio}
\usage{
sf_seas_ratio(df_original, df_dessaz, nyears = NULL)
}
\arguments{
\item{df_original}{Dataframe de entrada com a série original;}

\item{df_dessaz}{Dataframe de entrada com a série dessaz e a projeção já realizada até horizonte desejado;}

\item{nyears}{Opcional, número de anos que será utilizado do histórico para as computações. Se não preenchido, usa o histórico inteiro.}
}
\value{
Pode ser o mesmo df_original com a projeção ou apenas o histórico média do fator sazonal.
}
\description{
Realiza a projeção de uma série original usando a projeção de sua série dessaz,
fazendo a compatibilização pela razão histórica entre a série original e a dessaz.

Caso seja fornecido uma série dessaz sem projeção, o retorno passa a ser apenas o fator sazonal histórico médio.
}
\examples{
sf_seas_ratio(df_original = df_cleaned,
              df_dessaz = df_forecast_dessaz,
              nyears = 5)

}
\author{
Gabriel Bellé
}
