% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_snaive.R
\name{sf_snaive}
\alias{sf_snaive}
\title{Seasonal naïve method}
\usage{
sf_snaive(df, nyears = NULL, end_forecast)
}
\arguments{
\item{df}{DataFrame contendo a série limpa e organizada;}

\item{nyears}{Opcional, número de anos que será utilizado do histórico para as computações. Se não preenchido, usa o histórico inteiro.}

\item{end_forecast}{Chr representa a data final para a projeção;}
}
\value{
Retorna o mesmo df de input, porém com a projeção Seasonal naïve.
}
\description{
Método Seasonal Naïve, no qual cada previsão é igual ao último valor observado da mesma estação
(por exemplo, o mesmo mês do ano anterior) ou a média dos meses de X anos.
}
\details{
O @param df de entrada deve conter, pelo menos, as colunas de:
\code{date}: Data da observação:
\code{vl}: valor da observação.
}
\examples{
\dontrun{
sf_snaive(df, nyears = 1, end_forecast = "2026-12-01")

}

}
\author{
Luiz Paulo T.
}
