% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_topdown.R
\name{sf_topdown}
\alias{sf_topdown}
\title{Forecasting top down}
\usage{
sf_topdown(target_agg, ...)
}
\arguments{
\item{target_agg}{Dataframe contendo a série (top) já com projeção;}

\item{...}{Dataframe, N dfs que serão projetados utilizando a projeção contida em target_agg}
}
\value{
Retorna um dataframe com colunas de \code{date}, \code{forecast} e uma
coluna de valor para cada série que foi projetada, na mesma ordem em que foi inputada na função.
}
\description{
Realiza a projeção de N séries utilizando uma projeção já feita em outra série.
Utiliza o YoY da série que já possui projeção (top) para projetar as demais séries (down).
Por manter o mesmo YoY em cada mês para as séries (down), sempre haverá compatibilidade com a série (top).
Exp: ao projetar aberturas da PMC/PMS/PIM/PIB, não importa como as aberturas são compostas para chegar no indicador geral,
sempre estará compatibilizado.
}
\details{
O @param target_agg deve conter as seguintes colunas:
\code{date}: Data da observação:
\code{vl}: valor da observação;
\code{forecast}: bool indicando se o valor é projeção.

Os dfs passados em @param ... devem conter as colunas:
\code{date}: Data da observação:
\code{vl}: valor da observação e;

deve possuir mesma periodicidade que o @param target_agg.
}
\author{
Gabriel Bellé
}
