% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_series.R
\name{split_fs_series}
\alias{split_fs_series}
\alias{split_series}
\title{Split loaded series from FS}
\usage{
split_series(df, sid, type = "ambos")
}
\arguments{
\item{df}{Dataframe contendo a base de dados a ser quebrada;}

\item{sid}{Chr contendo o nome da série a ser retornada;}

\item{type}{Chr especificando tipo de retorno (realizado, projetado, ambos)}
}
\value{
O retorno é um df contendo os valores da série e a data.
}
\description{
Retorna apenas uma série de um df contendo séries empilhadas.
Pode ser útil como função auxiliar.
}
\details{
O input @param df deve ser um DataFrame contendo pelo menos as colunas de:
\code{sid}: SériesID da série;
\code{date}: Data da observação:
\code{forecast}: TRUE quando a observação for uma projeção;
\code{vl}: valor da observação.
}
\examples{
\dontrun{
split_fs_series(df = cleaned_df,
                sid = 'BREMP0018000OOML'),
                type = 'ambos')
}

}
\author{
Gabriel Bellé
}
