% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_series.R
\name{split_series}
\alias{split_series}
\title{Split loaded series from FS}
\usage{
split_series(df, sid_vl, estimate = NULL)
}
\arguments{
\item{df}{Dataframe contendo a base de dados a ser quebrada;}

\item{sid_vl}{Chr contendo o SeriesID da série a ser retornada;}

\item{estimate}{bool indicando se deve retornar dado realizado ou projetado. Se NULL retorna ambas.}
}
\value{
O retorno é um df contendo as colunas de date, vl, forecast.
}
\description{
Seleciona apenas uma série do retorno da fct load_clean_series.
}
\details{
Seleciona apenas um SeriesID (sid) do Dataframe de retorno da função load_clean_series.
}
\examples{
split_fs_series(df = cleaned_df,
                sid_vl = 'BREMP0018000OOML'),
                estimate = T)

}
\author{
Gabriel Bellé
}
