% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_periodicity.R
\name{get_periodicity}
\alias{get_periodicity}
\title{Discover periodicity from df}
\usage{
get_periodicity(df)
}
\arguments{
\item{df}{Dataframe contendo a base de dados.}
}
\value{
O retorno é uma lista contendo os parametros de:
\code{p_name}: nome da periodicidade;
\code{p_nmonths}: quantos meses há na periodicidade;
\code{p_ngap}: quantos meses pular até a próxima observação.
}
\description{
Descobre a periodicidade de um dataframe e cria parametros adequados a ela.
}
\details{
O input deve ser um df contendo pelo menos as colunas de:
\code{date}: Data da observação:
\code{vl}: valor da observação.
}
\examples{
\dontrun{
get_periodicity(df = cleaned_df)
}

}
\author{
Gabriel Bellé
}
