% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seas_adj.R
\name{get_seas_adj}
\alias{get_seas_adj}
\title{Seasonality adjust}
\usage{
get_seas_adj(df, type = "mean")
}
\arguments{
\item{df}{Dataframe contendo a série a ser dessazonalizada;}

\item{type}{Chr (STL, X13, mean) tipo de dessaz a ser aplicado.}
}
\value{
O retorno é um df contendo os valores da série dessazonalidada e a data.
}
\description{
Calcula o dessaz de uma série com as configurações automáticas
do STL, do X13, ou da média entre eles.
}
\details{
Chama as funções do pacote, get_stl e get_x13, onde de fato são feitas as estimações.
A função realiza dessaz tanto de série mensal quanto de série trimestral, sem necessidade de parâmetro extra.
}
\examples{
get_seas_adj(df = cleaned_df, type = 'mean')

}
\author{
Gabriel Bellé
}
