% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stl.R
\name{get_stl}
\alias{get_stl}
\title{Seasonal Decomposition of Time Series by Loess (STL)}
\usage{
get_stl(df)
}
\arguments{
\item{df}{representa a série para dessazonalização}
}
\value{
O retorno é um df contendo os valores da série dessazonalidada e a data.
}
\description{
Calcula o dessaz de uma série com as configurações automáticas
do STL.
}
\details{
O input deve ser um df contendo pelo as colunas de:
\code{date}: Data da observação:
\code{vl}: valor da observação.
}
\examples{
\dontrun{
get_stl(df = cleaned_df)
}

}
\author{
Gabriel Bellé
}
