% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_x13.R
\name{get_x13}
\alias{get_x13}
\title{X-13 ARIMA-SEATS}
\usage{
get_x13(df)
}
\arguments{
\item{df}{série para dessazonalização}
}
\value{
O retorno é um df contendo os valores da série dessazonalizada e a data.
}
\description{
Calcula o dessaz de uma série com as configurações automáticas
do X13.
}
\details{
O input deve ser um df contendo pelo as colunas de:
\code{date}: Data da observação:
\code{vl}: valor da observação.
}
\examples{
\dontrun{
get_x13(df = cleaned_df)
}

}
\author{
Gabriel Bellé
}
