% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clean_series.R
\name{load_clean_series}
\alias{load_clean_series}
\title{Load and Clean Feature Store (FS) series}
\usage{
load_clean_series(sids, auth_path, estimate = NULL)
}
\arguments{
\item{sids}{Vetor do tipo \code{chr} contendo as séries a serem retornadas da FS;}

\item{auth_path}{Chr contendo o caminho para o arquivo de autenticação do series.4macro}
}
\value{
O retorno é um df contendo as colunas:
\code{sid}: SeriesID da série;
\code{date}: Data da observação:
\code{forecast}: TRUE quando a observação for uma projeção;
\code{vl}: valor da observação.
}
\description{
Carrega e limpas as séries da FS
}
\details{
Baixa as séries da FS utilizados o pacote series.4macro e limpa elas
para que fiquem no formato de \code{tibble}, empiladas e indicando onde é a projeção.
}
\examples{
\dontrun{
load_clean_series(sids = c('BREMP0018000OOML', 'BREMP0018000SOML'),
                  auth_path = 'User/auth.ini')
}

}
\author{
Gabriel Bellé
}
