% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_hw.R
\name{sf_hw}
\alias{sf_hw}
\title{Holt-Winters}
\usage{
sf_hw(df, type = NULL, end_forecast)
}
\arguments{
\item{df}{DataFrame contendo a série limpa e organizada;}

\item{type}{Tipo de Holt-Winters: mutiplicativo, aditivo, tendência e NULL (default);}

\item{end_forecast}{representa a data final, último mês ou trimestre, da projeção.}
}
\value{
Retorna o mesmo df de input, porém a projeção formalmente aplicada com o método Holt-Winters
}
\description{
Aplicação do método Holt-Winters para projeção
}
\details{
As opções de type são multiplicative, additive, trend e NULL. O quais representam, respectivamente: sazonalidade mutiplicativa e aditiva, o modelo Holt-Winters com tendência e, por fim, NULL representa o default, isto é, a escolha automática via critério de informação.

O input deve ser um data.frame limpo e organizado:
\code{date}: Data da observação:
\code{vl}: valor da observação.
}
\examples{
\dontrun{
sf_hw(df, type = NULL, end_forecast = "2026-12-01")

}

}
\author{
Luiz Paulo T.
}
