% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_topdown.R
\name{sf_topdown}
\alias{sf_topdown}
\title{Top-down forecasting}
\usage{
sf_topdown(target_agg, ...)
}
\arguments{
\item{target_agg}{Dataframe contendo a série (top) já com projeção;}

\item{...}{Dataframe, N dfs que serão projetados utilizando a projeção contida em target_agg}
}
\value{
Retorna um dataframe com colunas de \code{date}, \code{forecast} e uma
coluna de valor para cada série que foi projetada, na mesma ordem em que foi inputada na função.
}
\description{
Realiza a projeção de N séries a partir de uma projeção já feita em outra série.
Utiliza o YoY da série que já possui projeção (top) para projetar as demais séries (down).
}
\details{
Pelo método manter o mesmo YoY em cada mês para as séries (down), sempre haverá compatibilidade com a série (top).

Exp: ao projetar aberturas da PMC/PMS/PIM/PIB, não importa como as aberturas são compostas para chegar no indicador geral,
sempre estará compatibilizado. Note-se que as aberturas terão o mesmo comportamento que a abertura geral.

Os Dfs devem possuir mesma periodicidade que o target_agg.
}
\examples{
sf_topdown(target_agg = pmc_geral,
           pmc_construcao, pmc_eletrodomestico, pmc_vestuario)

}
\author{
Gabriel Bellé
}
